"acdc-archive" class

An S3 class that defines the object returned by an acoustic-container/depth-contour (AC/DC) algorithm (ac, dc or acdc).

Value

An acdc_archive-class object is a named list that contains the output of a call to ac, dc or acdc. This contains the following elements: `archive', `ts_by_chunk', `time' and `args'.

The main output is the `archive' element. This contains a list of acdc_record-class objects returned by the call(s) to the internal routines that implement the algorithm. If the algorithm is implemented step-wise, this list contains a single element; if the algorithm is implemented chunk-wise, there is one element for each chunk. Each element contains an acdc_record-class object with results of the call to internal routines for that chunk. The results across chunks can be aggregated using acdc_simplify.

The `ts_by_chunk' element is a list, with one element for each chunk, that contains the movement time series that were used in that chunk. Each element contains `acoustics' and `archival' dataframes, though the former is NULL for the DC algorithm and the latter is NULL for the AC algorithm. For the AC* algorithms, if there is more than one chunk, the last observation of each acoustic chunk is the same as the first acoustic observation for the next chunk. This results from splitting chunks at acoustic observations, which enables the results of chunks that are computed independently to be simply aggregated without the loss of any information.

The `time' element is a dataframe that defines the times of sequential stages in the algorithm's progression, providing a record of computation time.

The `args' element is a named list of user inputs that records the parameters used to generate the outputs.

Author

Edward Lavender