Example DCPF algorithm output: particle histories

An example pf_archive-class object. This comprises a named list of results from an implementation of the depth-contour particle filtering (DCPF) algorithm (see dc and pf). The algorithm was implemented for a series of depth observations collected from flapper skate (Dipturus intermedius) off the west coast of Scotland, where the bathymetry is described by dat_gebco. The outputs include algorithm parameters and a list of dataframes that record the particles sampled at each time step. This is included principally to streamline function examples.

dat_dcpf_histories

Format

A named list, with two elements, which records the parameters used in the call to pf and the particles sampled at each time step:

args

A named list that records the function arguments used to generate outputs.

history

A list of dataframes, one for each time step, that record particle samples. Each dataframe comprises n = 10 rows and the following three columns:

id_current An integer that uniquely defines each location on the bathymetry raster (bathy). pr_current A double that defines the probability of movement into each cell. timestep An integer that defines each time step.