Example passive acoustic telemetry receiver moorings dataset

A dataset containing a sample of passive acoustic telemetry receiver locations and associated information. Data are arranged by receiver_id (see below).

dat_moorings

Format

A dataframe with 40 observations and 8 variables:

receiver_id

An integer which distinguishes each unique receiver deployment.

receiver

A character which defines each receiver.

sentinel_id

A character which defines each receiver's built-in acoustic transmitter (`sentinel tag'). Some receivers did not have sentinel tags; their sentinel_id is NA.

receiver_start_date

A date which defines the start date of each receiver's deployment.

receiver_end_date

A date which defines the end date of each receiver's deployment.

receiver_long

A number which defines the longitude (decimal degrees) of each receiver.

receiver_lat

A number which defines the latitude (decimal degrees) of each receiver.

receiver_depth

A number which defines the approximate depth of each receiver below the surface (m).

Source

Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.

References

Data collection and processing is described in Lavender (in prep). Fine-scale habitat use of the Critically Endangered flapper skate (Dipturus intermedius). [Doctoral dissertation, University of St Andrews].