Compute Euclidean distances between receivers

This function computes Euclidean distances (km) between all combinations of receivers.

dist_btw_receivers(moorings, f = NULL)

Arguments

moorings
A dataframe which defines each unique receiver deployment. This should contain the columns: `receiver_id', a unique identifier of each receiver, `receiver_lat', the latitude of that receiver in decimal degrees; and `receiver_long', the longitude of that receiver in decimal degrees (see dat_moorings).

f
(optional) A function which is used to process distances before these are returned. For example, it may be useful to round distances to nearest km with f = function(x) round(x, digits = 0).

Value

The function returns a dataframe with columns `r1', `r2' and `dist'. These define the IDs of each combination of receivers and the associated distance between them, in km. Note that the dataframe contains duplicate combinations of receivers (e.g., both r1 = 1 and r2 = 2 and r1 = 2 and r2 = 1).

Author

Edward Lavender

Examples

#### Example (1): Compute distances between all combinations of receivers in km
# Define dataframe with required columns
dat <- data.frame(receiver_id = dat_moorings$receiver_id,
                  receiver_lat = dat_moorings$receiver_lat,
                  receiver_long = dat_moorings$receiver_long)
# Compute distances
dist_btw_receivers_km <- dist_btw_receivers(dat)
head(dist_btw_receivers_km)

#>   r1 r2      dist
#> 1  3  3  0.000000
#> 2  4  3 14.111344
#> 3  7  3 15.308019
#> 4  9  3  0.542825
#> 5 11  3 13.765477
#> 6 12  3 13.796536

#### Example (2): Post-process distances via the f argument
dist_btw_receivers_km_round <- dist_btw_receivers(dat, f = round)
head(dist_btw_receivers_km_round)

#>   r1 r2 dist
#> 1  3  3    0
#> 2  4  3   14
#> 3  7  3   15
#> 4  9  3    1
#> 5 11  3   14
#> 6 12  3   14

# convert distances to m
dist_btw_receivers_m <- dist_btw_receivers(dat, f = function(x) x*1000)
head(dist_btw_receivers_m)

#>   r1 r2      dist
#> 1  3  3     0.000
#> 2  4  3 14111.344
#> 3  7  3 15308.019
#> 4  9  3   542.825
#> 5 11  3 13765.477
#> 6 12  3 13796.536