Global example controls

These variables provide convenient switches for turning examples on/off during package testing. They can be overridden by the user.

flapper_run_parallel

flapper_run_slow

Format

An object of class logical of length 1.

An object of class logical of length 1.

Details

A set of logical variables:

flapper_run_parallel is a logical variable that defines whether or not to run selected parallel examples.

flapper_run_slow is logical variable that defines whether or not to run selected `slow' examples.

Author

Edward Lavender

Examples

#### Example (1): flapper_run_parallel
## Illustration of intended usage for a parallelised function
if(flapper_run_parallel){
  cl_lapply(1:10,
            function(x) x + 1,
            cl = parallel::makeCluster(2L))
}

#> [[1]]
#> [1] 2
#> 
#> [[2]]
#> [1] 3
#> 
#> [[3]]
#> [1] 4
#> 
#> [[4]]
#> [1] 5
#> 
#> [[5]]
#> [1] 6
#> 
#> [[6]]
#> [1] 7
#> 
#> [[7]]
#> [1] 8
#> 
#> [[8]]
#> [1] 9
#> 
#> [[9]]
#> [1] 10
#> 
#> [[10]]
#> [1] 11
#> 

#### Example (2): flapper_run_slow
## Illustration of intended usage for a slow function
if(flapper_run_slow){
  pf_args               <- dat_dcpf_histories$args
  pf_args$n             <- 100
  pf_args$calc_distance <- "lcp"
  pf_args$seed          <- 1
  out_pf                <- do.call(pf, pf_args)
}

#> flapper::pf() called (@ 2022-05-28 18:09:37)... 
#> ... Setting up function... 
#> ... Setting up cost-surface for calc_distance = 'lcp'... 
#> flapper::lcp_costs() called (@ 2022-05-28 18:09:37)... 
#> ... Defining transition matrices... 

#> Warning: transition function gives negative values

#> ... Calculating distance matrices... 
#> ... Assembling LCP costs... 
#> ... flapper::lcp_costs() call completed (@ 2022-05-28 18:09:37) after ~0 minutes. 
#> flapper::lcp_graph_surface() called (@ 2022-05-28 18:09:37)... 
#> ... Defining nodes, edges and costs to make graph... 
#> ... Constructing graph object... 
#> ... flapper::lcp_graph_surface() call completed (@ 2022-05-28 18:09:37) after ~0 minutes. 
#> ... Determining the set of possible starting locations (t = 1)... 
#> flapper::lcp_from_point() called (@ 2022-05-28 18:09:37)... 
#> ... Calling cppRouting::get_distance_matrix() to get least-cost distances... 
#> ... Making raster of least-cost distances from origin ... 
#> ... flapper::lcp_from_point() call completed (@ 2022-05-28 18:09:37) after ~0 minutes. 
#> ... Implementing algorithm iteratively over time steps... 
#> ... ... Time = 1... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 2... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 3... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 4... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 5... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 6... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 7... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 8... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 9... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 10... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 11... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 12... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 13... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 14... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 15... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 16... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 17... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 18... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... flapper::pf() call completed (@ 2022-05-28 18:10:12) after ~0.58 minutes.