Short-cuts to elements of an "acdc_record" object

These functions provide short-cuts to elements of an acdc_record-class object.

acdc_access_dat(record)

acdc_access_timesteps(record)

acdc_access_maps(
  record,
  type = c("map_timestep", "map_cumulative"),
  select = NULL
)

Arguments

record
An acdc_record-class object.

type
For acdc_access_maps, type is a character that specifies whether or not to access time step-specific maps (type = "map_timestep") or cumulative maps (type = "map_cumulative").

select
(optional) For acdc_access_maps, select is an integer vector that defines the cumulative time steps (i.e., accounting for both primary (acoustic) and secondary (archival) time steps) for which maps are required.

Value

acdc_access_dat returns a single dataframe for all time steps.

acdc_access_timesteps returns an integer that defines the total number of time steps.

acdc_access_maps returns a single list of time-step specific or cumulative maps for specified or all time steps.

Details

acdc_access_dat accesses the record$dat elements of an acdc_record-class object for all time steps.

acdc_access_timesteps accesses the total number of time steps stored in an acdc_record-class object, accounting for both primary (acoustic) and secondary (archival) time steps.

acdc_access_maps accesses all, or specified, maps from the record$spatial elements of an acdc_record-class object.

Author

Edward Lavender

Examples

#### Example (1): acdc_access_dat()
acdc_access_dat(acdc_simplify(dat_acdc))

#> acdc_simplify() implemented for type = 'acs'.

#>     timestep_cumulative timestep_detection timestep_archival receiver_1_id
#> 1                     1                  1                 1            26
#> 2                     2                  2                 1            26
#> 3                     3                  3                 1            26
#> 4                     4                  3                 2            26
#> 5                     5                  4                 1            26
#> 6                     6                  5                 1            26
#> 7                     7                  5                 2            26
#> 8                     8                  6                 1            26
#> 9                     9                  7                 1            26
#> 10                   10                  8                 1            26
#> 11                   11                  9                 1            26
#> 12                   12                 10                 1            26
#> 13                   13                 11                 1            26
#> 14                   14                 12                 1            26
#> 15                   15                 13                 1            26
#> 16                   16                 14                 1            26
#> 17                   17                 15                 1            26
#> 18                   18                 16                 1            26
#> 19                   19                 16                 2            26
#> 20                   20                 16                 3            26
#> 21                   21                 17                 1            26
#> 22                   22                 18                 1            26
#> 23                   23                 18                 2            26
#> 24                   24                 18                 3            26
#> 25                   25                 18                 4            26
#> 26                   26                 18                 5            26
#> 27                   27                 18                 6            26
#> 28                   28                 18                 7            26
#> 29                   29                 18                 8            26
#> 30                   30                 18                 9            26
#> 31                   31                 18                10            26
#> 32                   32                 19                 1            21
#> 33                   33                 20                 1            26
#> 34                   34                 21                 1            26
#> 35                   35                 22                 1            21
#> 36                   36                 23                 1            21
#> 37                   37                 24                 1            21
#> 38                   38                 25                 1            21
#> 39                   39                 26                 1            21
#> 40                   40                 27                 1            21
#> 41                   41                 28                 1            21
#> 42                   42                 29                 1            21
#> 43                   43                 30                 1            21
#> 44                   44                 30                 2            21
#> 45                   45                 31                 1            21
#> 46                   46                 32                 1            21
#> 47                   47                 33                 1            21
#> 48                   48                 34                 1            21
#> 49                   49                 35                 1            21
#> 50                   50                 36                 1            21
#> 51                   51                 37                 1            21
#> 52                   52                 38                 1            21
#> 53                   53                 39                 1            21
#> 54                   54                 39                 2            21
#> 55                   55                 40                 1            21
#> 56                   56                 41                 1            21
#> 57                   57                 42                 1            21
#> 58                   58                 43                 1            26
#> 59                   59                 44                 1            26
#> 60                   60                 45                 1            26
#> 61                   61                 46                 1            26
#> 62                   62                 47                 1            26
#> 63                   63                 47                 2            26
#> 64                   64                 48                 1            26
#> 65                   65                 48                 2            26
#> 66                   66                 48                 3            26
#> 67                   67                 48                 4            26
#> 68                   68                 48                 5            26
#> 69                   69                 49                 1            26
#> 70                   70                 50                 1            26
#> 71                   71                 51                 1            26
#> 72                   72                 52                 1            21
#> 73                   73                 52                 2            21
#> 74                   74                 53                 1            21
#> 75                   75                 53                 2            21
#> 76                   76                 53                 3            21
#> 77                   77                 53                 4            21
#> 78                   78                 53                 5            21
#> 79                   79                 53                 6            21
#> 80                   80                 53                 7            21
#> 81                   81                 53                 8            21
#> 82                   82                 53                 9            21
#> 83                   83                 53                10            21
#> 84                   84                 54                 1            26
#> 85                   85                 55                 1            26
#> 86                   86                 55                 2            26
#> 87                   87                 56                 1            26
#> 88                   88                 56                 2            26
#> 89                   89                 56                 3            26
#> 90                   90                 57                 1            26
#> 91                   91                 57                 2            26
#> 92                   92                 58                 1            26
#> 93                   93                 59                 1            26
#> 94                   94                 59                 2            26
#> 95                   95                 60                 1            26
#> 96                   96                 61                 1            26
#> 97                   97                 62                 1            26
#> 98                   98                 63                 1            26
#> 99                   99                 64                 1            26
#> 100                 100                 65                 1            26
#> 101                 101                 66                 1            26
#> 102                 102                 67                 1            31
#> 103                 103                 68                 1            26
#> 104                 104                 69                 1            31
#> 105                 105                 70                 1            26
#> 106                 106                 71                 1            26
#> 107                 107                 72                 1            26
#> 108                 108                 73                 1            26
#> 109                 109                 74                 1            26
#> 110                 110                 75                 1            31
#> 111                 111                 76                 1            26
#> 112                 112                 77                 1            26
#> 113                 113                 78                 1            26
#> 114                 114                 79                 1            26
#> 115                 115                 80                 1            26
#> 116                 116                 81                 1            26
#> 117                 117                 82                 1            26
#> 118                 118                 83                 1            26
#> 119                 119                 84                 1            26
#> 120                 120                 85                 1            26
#> 121                 121                 86                 1            26
#> 122                 122                 86                 2            26
#> 123                 123                 86                 3            26
#> 124                 124                 86                 4            26
#> 125                 125                 86                 5            26
#> 126                 126                 86                 6            26
#> 127                 127                 86                 7            26
#> 128                 128                 86                 8            26
#> 129                 129                 86                 9            26
#> 130                 130                 86                10            26
#> 131                 131                 86                11            26
#> 132                 132                 86                12            26
#> 133                 133                 86                13            26
#> 134                 134                 86                14            26
#> 135                 135                 86                15            26
#> 136                 136                 86                16            26
#> 137                 137                 86                17            26
#> 138                 138                 86                18            26
#> 139                 139                 86                19            26
#> 140                 140                 86                20            26
#> 141                 141                 86                21            26
#> 142                 142                 86                22            26
#> 143                 143                 86                23            26
#> 144                 144                 86                24            26
#> 145                 145                 86                25            26
#> 146                 146                 86                26            26
#> 147                 147                 86                27            26
#> 148                 148                 86                28            26
#> 149                 149                 86                29            26
#> 150                 150                 86                30            26
#> 151                 151                 86                31            26
#> 152                 152                 86                32            26
#> 153                 153                 86                33            26
#> 154                 154                 86                34            26
#> 155                 155                 86                35            26
#> 156                 156                 86                36            26
#> 157                 157                 86                37            26
#> 158                 158                 86                38            26
#> 159                 159                 86                39            26
#> 160                 160                 86                40            26
#> 161                 161                 86                41            26
#> 162                 162                 86                42            26
#> 163                 163                 86                43            26
#> 164                 164                 86                44            26
#> 165                 165                 86                45            26
#> 166                 166                 86                46            26
#> 167                 167                 86                47            26
#> 168                 168                 86                48            26
#> 169                 169                 86                49            26
#> 170                 170                 86                50            26
#> 171                 171                 86                51            26
#> 172                 172                 86                52            26
#> 173                 173                 86                53            26
#> 174                 174                 86                54            26
#> 175                 175                 86                55            26
#> 176                 176                 86                56            26
#> 177                 177                 86                57            26
#> 178                 178                 86                58            26
#> 179                 179                 86                59            26
#> 180                 180                 86                60            26
#> 181                 181                 86                61            26
#> 182                 182                 86                62            26
#> 183                 183                 86                63            26
#> 184                 184                 86                64            26
#> 185                 185                 86                65            26
#> 186                 186                 86                66            26
#> 187                 187                 86                67            26
#> 188                 188                 86                68            26
#> 189                 189                 86                69            26
#> 190                 190                 86                70            26
#> 191                 191                 86                71            26
#> 192                 192                 86                72            26
#> 193                 193                 86                73            26
#> 194                 194                 86                74            26
#> 195                 195                 86                75            26
#> 196                 196                 86                76            26
#> 197                 197                 86                77            26
#> 198                 198                 86                78            26
#> 199                 199                 86                79            26
#> 200                 200                 86                80            26
#> 201                 201                 86                81            26
#> 202                 202                 86                82            26
#> 203                 203                 86                83            26
#> 204                 204                 86                84            26
#> 205                 205                 86                85            26
#> 206                 206                 86                86            26
#> 207                 207                 86                87            26
#> 208                 208                 86                88            26
#> 209                 209                 87                 1            26
#> 210                 210                 87                 2            26
#> 211                 211                 88                 1            26
#> 212                 212                 89                 1            26
#> 213                 213                 90                 1            26
#> 214                 214                 91                 1            26
#> 215                 215                 91                 2            26
#> 216                 216                 92                 1            26
#> 217                 217                 93                 1            31
#> 218                 218                 94                 1            26
#> 219                 219                 95                 1            31
#> 220                 220                 96                 1            31
#> 221                 221                 97                 1            26
#> 222                 222                 97                 2            26
#> 223                 223                 98                 1            31
#> 224                 224                 99                 1            26
#> 225                 225                100                 1            26
#> 226                 226                101                 1            31
#>     receiver_2_id receiver_1_timestamp receiver_2_timestamp time_btw_dets
#> 1              26  2016-03-17 01:50:00  2016-03-17 01:52:00           120
#> 2              26  2016-03-17 01:52:00  2016-03-17 01:54:00           120
#> 3              26  2016-03-17 01:54:00  2016-03-17 01:58:00           240
#> 4              26  2016-03-17 01:54:00  2016-03-17 01:58:00           240
#> 5              26  2016-03-17 01:58:00  2016-03-17 02:00:00           120
#> 6              26  2016-03-17 02:00:00  2016-03-17 02:04:00           240
#> 7              26  2016-03-17 02:00:00  2016-03-17 02:04:00           240
#> 8              26  2016-03-17 02:04:00  2016-03-17 02:06:00           120
#> 9              26  2016-03-17 02:06:00  2016-03-17 02:08:00           120
#> 10             26  2016-03-17 02:08:00  2016-03-17 02:10:00           120
#> 11             26  2016-03-17 02:10:00  2016-03-17 02:12:00           120
#> 12             26  2016-03-17 02:12:00  2016-03-17 02:14:00           120
#> 13             26  2016-03-17 02:14:00  2016-03-17 02:16:00           120
#> 14             26  2016-03-17 02:16:00  2016-03-17 02:18:00           120
#> 15             26  2016-03-17 02:18:00  2016-03-17 02:20:00           120
#> 16             26  2016-03-17 02:20:00  2016-03-17 02:22:00           120
#> 17             26  2016-03-17 02:22:00  2016-03-17 02:24:00           120
#> 18             26  2016-03-17 02:24:00  2016-03-17 02:30:00           360
#> 19             26  2016-03-17 02:24:00  2016-03-17 02:30:00           360
#> 20             26  2016-03-17 02:24:00  2016-03-17 02:30:00           360
#> 21             26  2016-03-17 02:30:00  2016-03-17 02:32:00           120
#> 22             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 23             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 24             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 25             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 26             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 27             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 28             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 29             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 30             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 31             21  2016-03-17 02:32:00  2016-03-17 02:52:00          1200
#> 32             26  2016-03-17 02:52:00  2016-03-17 02:54:00           120
#> 33             26  2016-03-17 02:54:00  2016-03-17 02:56:00           120
#> 34             21  2016-03-17 02:56:00  2016-03-17 02:56:00             0
#> 35             21  2016-03-17 02:56:00  2016-03-17 02:58:00           120
#> 36             21  2016-03-17 02:58:00  2016-03-17 03:00:00           120
#> 37             21  2016-03-17 03:00:00  2016-03-17 03:02:00           120
#> 38             21  2016-03-17 03:02:00  2016-03-17 03:04:00           120
#> 39             21  2016-03-17 03:04:00  2016-03-17 03:06:00           120
#> 40             21  2016-03-17 03:06:00  2016-03-17 03:08:00           120
#> 41             21  2016-03-17 03:08:00  2016-03-17 03:10:00           120
#> 42             21  2016-03-17 03:10:00  2016-03-17 03:12:00           120
#> 43             21  2016-03-17 03:12:00  2016-03-17 03:16:00           240
#> 44             21  2016-03-17 03:12:00  2016-03-17 03:16:00           240
#> 45             21  2016-03-17 03:16:00  2016-03-17 03:18:00           120
#> 46             21  2016-03-17 03:18:00  2016-03-17 03:20:00           120
#> 47             21  2016-03-17 03:20:00  2016-03-17 03:22:00           120
#> 48             21  2016-03-17 03:22:00  2016-03-17 03:24:00           120
#> 49             21  2016-03-17 03:24:00  2016-03-17 03:26:00           120
#> 50             21  2016-03-17 03:26:00  2016-03-17 03:28:00           120
#> 51             21  2016-03-17 03:28:00  2016-03-17 03:30:00           120
#> 52             21  2016-03-17 03:30:00  2016-03-17 03:32:00           120
#> 53             21  2016-03-17 03:32:00  2016-03-17 03:36:00           240
#> 54             21  2016-03-17 03:32:00  2016-03-17 03:36:00           240
#> 55             21  2016-03-17 03:36:00  2016-03-17 03:38:00           120
#> 56             21  2016-03-17 03:38:00  2016-03-17 03:40:00           120
#> 57             26  2016-03-17 03:40:00  2016-03-17 03:42:00           120
#> 58             26  2016-03-17 03:42:00  2016-03-17 03:44:00           120
#> 59             26  2016-03-17 03:44:00  2016-03-17 03:46:00           120
#> 60             26  2016-03-17 03:46:00  2016-03-17 03:48:00           120
#> 61             26  2016-03-17 03:48:00  2016-03-17 03:50:00           120
#> 62             26  2016-03-17 03:50:00  2016-03-17 03:54:00           240
#> 63             26  2016-03-17 03:50:00  2016-03-17 03:54:00           240
#> 64             26  2016-03-17 03:54:00  2016-03-17 04:04:00           600
#> 65             26  2016-03-17 03:54:00  2016-03-17 04:04:00           600
#> 66             26  2016-03-17 03:54:00  2016-03-17 04:04:00           600
#> 67             26  2016-03-17 03:54:00  2016-03-17 04:04:00           600
#> 68             26  2016-03-17 03:54:00  2016-03-17 04:04:00           600
#> 69             26  2016-03-17 04:04:00  2016-03-17 04:06:00           120
#> 70             26  2016-03-17 04:06:00  2016-03-17 04:08:00           120
#> 71             21  2016-03-17 04:08:00  2016-03-17 04:08:00             0
#> 72             21  2016-03-17 04:08:00  2016-03-17 04:12:00           240
#> 73             21  2016-03-17 04:08:00  2016-03-17 04:12:00           240
#> 74             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 75             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 76             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 77             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 78             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 79             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 80             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 81             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 82             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 83             26  2016-03-17 04:12:00  2016-03-17 04:32:00          1200
#> 84             26  2016-03-17 04:32:00  2016-03-17 04:34:00           120
#> 85             26  2016-03-17 04:34:00  2016-03-17 04:38:00           240
#> 86             26  2016-03-17 04:34:00  2016-03-17 04:38:00           240
#> 87             26  2016-03-17 04:38:00  2016-03-17 04:44:00           360
#> 88             26  2016-03-17 04:38:00  2016-03-17 04:44:00           360
#> 89             26  2016-03-17 04:38:00  2016-03-17 04:44:00           360
#> 90             26  2016-03-17 04:44:00  2016-03-17 04:48:00           240
#> 91             26  2016-03-17 04:44:00  2016-03-17 04:48:00           240
#> 92             26  2016-03-17 04:48:00  2016-03-17 04:50:00           120
#> 93             26  2016-03-17 04:50:00  2016-03-17 04:54:00           240
#> 94             26  2016-03-17 04:50:00  2016-03-17 04:54:00           240
#> 95             26  2016-03-17 04:54:00  2016-03-17 04:56:00           120
#> 96             26  2016-03-17 04:56:00  2016-03-17 04:58:00           120
#> 97             26  2016-03-17 04:58:00  2016-03-17 05:00:00           120
#> 98             26  2016-03-17 05:00:00  2016-03-17 05:02:00           120
#> 99             26  2016-03-17 05:02:00  2016-03-17 05:04:00           120
#> 100            26  2016-03-17 05:04:00  2016-03-17 05:06:00           120
#> 101            31  2016-03-17 05:06:00  2016-03-17 05:06:00             0
#> 102            26  2016-03-17 05:06:00  2016-03-17 05:08:00           120
#> 103            31  2016-03-17 05:08:00  2016-03-17 05:10:00           120
#> 104            26  2016-03-17 05:10:00  2016-03-17 05:10:00             0
#> 105            26  2016-03-17 05:10:00  2016-03-17 05:12:00           120
#> 106            26  2016-03-17 05:12:00  2016-03-17 05:14:00           120
#> 107            26  2016-03-17 05:14:00  2016-03-17 05:16:00           120
#> 108            26  2016-03-17 05:16:00  2016-03-17 05:18:00           120
#> 109            31  2016-03-17 05:18:00  2016-03-17 05:20:00           120
#> 110            26  2016-03-17 05:20:00  2016-03-17 05:20:00             0
#> 111            26  2016-03-17 05:20:00  2016-03-17 05:22:00           120
#> 112            26  2016-03-17 05:22:00  2016-03-17 05:24:00           120
#> 113            26  2016-03-17 05:24:00  2016-03-17 05:26:00           120
#> 114            26  2016-03-17 05:26:00  2016-03-17 05:28:00           120
#> 115            26  2016-03-17 05:28:00  2016-03-17 05:30:00           120
#> 116            26  2016-03-17 05:30:00  2016-03-17 05:32:00           120
#> 117            26  2016-03-17 05:32:00  2016-03-17 05:34:00           120
#> 118            26  2016-03-17 05:34:00  2016-03-17 05:36:00           120
#> 119            26  2016-03-17 05:36:00  2016-03-17 05:38:00           120
#> 120            26  2016-03-17 05:38:00  2016-03-17 05:40:00           120
#> 121            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 122            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 123            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 124            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 125            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 126            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 127            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 128            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 129            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 130            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 131            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 132            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 133            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 134            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 135            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 136            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 137            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 138            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 139            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 140            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 141            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 142            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 143            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 144            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 145            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 146            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 147            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 148            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 149            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 150            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 151            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 152            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 153            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 154            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 155            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 156            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 157            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 158            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 159            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 160            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 161            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 162            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 163            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 164            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 165            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 166            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 167            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 168            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 169            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 170            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 171            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 172            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 173            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 174            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 175            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 176            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 177            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 178            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 179            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 180            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 181            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 182            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 183            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 184            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 185            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 186            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 187            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 188            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 189            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 190            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 191            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 192            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 193            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 194            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 195            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 196            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 197            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 198            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 199            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 200            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 201            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 202            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 203            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 204            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 205            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 206            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 207            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 208            26  2016-03-17 05:40:00  2016-03-17 08:36:00         10560
#> 209            26  2016-03-17 08:36:00  2016-03-17 08:40:00           240
#> 210            26  2016-03-17 08:36:00  2016-03-17 08:40:00           240
#> 211            26  2016-03-17 08:40:00  2016-03-17 08:42:00           120
#> 212            26  2016-03-17 08:42:00  2016-03-17 08:44:00           120
#> 213            26  2016-03-17 08:44:00  2016-03-17 08:46:00           120
#> 214            26  2016-03-17 08:46:00  2016-03-17 08:50:00           240
#> 215            26  2016-03-17 08:46:00  2016-03-17 08:50:00           240
#> 216            31  2016-03-17 08:50:00  2016-03-17 08:50:00             0
#> 217            26  2016-03-17 08:50:00  2016-03-17 08:52:00           120
#> 218            31  2016-03-17 08:52:00  2016-03-17 08:52:00             0
#> 219            31  2016-03-17 08:52:00  2016-03-17 08:54:00           120
#> 220            26  2016-03-17 08:54:00  2016-03-17 08:54:00             0
#> 221            31  2016-03-17 08:54:00  2016-03-17 08:58:00           240
#> 222            31  2016-03-17 08:54:00  2016-03-17 08:58:00           240
#> 223            26  2016-03-17 08:58:00  2016-03-17 08:58:00             0
#> 224            26  2016-03-17 08:58:00  2016-03-17 09:00:00           120
#> 225            31  2016-03-17 09:00:00  2016-03-17 09:00:00             0
#> 226            26  2016-03-17 09:00:00  2016-03-17 09:02:00           120
#>      archival_timestamp archival_depth chunk
#> 1   2016-03-17 01:50:00          73.78     1
#> 2   2016-03-17 01:52:00          73.32     1
#> 3   2016-03-17 01:54:00          73.32     1
#> 4   2016-03-17 01:56:00          73.32     1
#> 5   2016-03-17 01:58:00          73.55     1
#> 6   2016-03-17 02:00:00          68.70     1
#> 7   2016-03-17 02:02:00          72.16     1
#> 8   2016-03-17 02:04:00          72.85     1
#> 9   2016-03-17 02:06:00          73.55     1
#> 10  2016-03-17 02:08:00          73.32     1
#> 11  2016-03-17 02:10:00          73.32     1
#> 12  2016-03-17 02:12:00          73.55     1
#> 13  2016-03-17 02:14:00          73.32     1
#> 14  2016-03-17 02:16:00          73.32     1
#> 15  2016-03-17 02:18:00          73.32     1
#> 16  2016-03-17 02:20:00          72.39     1
#> 17  2016-03-17 02:22:00          72.39     1
#> 18  2016-03-17 02:24:00          72.39     1
#> 19  2016-03-17 02:26:00          72.39     1
#> 20  2016-03-17 02:28:00          72.39     1
#> 21  2016-03-17 02:30:00          72.39     1
#> 22  2016-03-17 02:32:00          65.23     1
#> 23  2016-03-17 02:34:00          73.55     1
#> 24  2016-03-17 02:36:00          75.16     1
#> 25  2016-03-17 02:38:00          76.78     1
#> 26  2016-03-17 02:40:00          76.55     1
#> 27  2016-03-17 02:42:00          76.78     1
#> 28  2016-03-17 02:44:00          76.55     1
#> 29  2016-03-17 02:46:00          71.01     1
#> 30  2016-03-17 02:48:00          72.16     1
#> 31  2016-03-17 02:50:00          72.16     1
#> 32  2016-03-17 02:52:00          73.78     1
#> 33  2016-03-17 02:54:00          71.70     1
#> 34  2016-03-17 02:56:00          65.92     1
#> 35  2016-03-17 02:56:00          65.92     1
#> 36  2016-03-17 02:58:00          66.15     1
#> 37  2016-03-17 03:00:00          66.15     1
#> 38  2016-03-17 03:02:00          66.39     1
#> 39  2016-03-17 03:04:00          66.39     1
#> 40  2016-03-17 03:06:00          66.15     1
#> 41  2016-03-17 03:08:00          66.15     1
#> 42  2016-03-17 03:10:00          66.15     1
#> 43  2016-03-17 03:12:00          66.15     1
#> 44  2016-03-17 03:14:00          66.15     1
#> 45  2016-03-17 03:16:00          66.15     1
#> 46  2016-03-17 03:18:00          66.39     1
#> 47  2016-03-17 03:20:00          66.15     2
#> 48  2016-03-17 03:22:00          66.15     2
#> 49  2016-03-17 03:24:00          66.39     2
#> 50  2016-03-17 03:26:00          66.15     2
#> 51  2016-03-17 03:28:00          65.92     2
#> 52  2016-03-17 03:30:00          66.15     2
#> 53  2016-03-17 03:32:00          66.15     2
#> 54  2016-03-17 03:34:00          66.39     2
#> 55  2016-03-17 03:36:00          66.15     2
#> 56  2016-03-17 03:38:00          66.15     2
#> 57  2016-03-17 03:40:00          65.92     2
#> 58  2016-03-17 03:42:00          65.69     2
#> 59  2016-03-17 03:44:00          22.77     2
#> 60  2016-03-17 03:46:00          49.99     2
#> 61  2016-03-17 03:48:00          47.45     2
#> 62  2016-03-17 03:50:00          49.76     2
#> 63  2016-03-17 03:52:00          52.30     2
#> 64  2016-03-17 03:54:00          55.76     2
#> 65  2016-03-17 03:56:00          65.46     2
#> 66  2016-03-17 03:58:00          68.23     2
#> 67  2016-03-17 04:00:00          70.08     2
#> 68  2016-03-17 04:02:00          71.70     2
#> 69  2016-03-17 04:04:00          60.15     2
#> 70  2016-03-17 04:06:00          42.84     2
#> 71  2016-03-17 04:08:00          38.69     2
#> 72  2016-03-17 04:08:00          38.69     2
#> 73  2016-03-17 04:10:00          40.76     2
#> 74  2016-03-17 04:12:00          39.38     2
#> 75  2016-03-17 04:14:00          36.38     2
#> 76  2016-03-17 04:16:00          35.00     2
#> 77  2016-03-17 04:18:00          39.61     2
#> 78  2016-03-17 04:20:00          43.30     2
#> 79  2016-03-17 04:22:00          43.78     2
#> 80  2016-03-17 04:24:00          40.30     2
#> 81  2016-03-17 04:26:00          43.76     2
#> 82  2016-03-17 04:28:00          43.53     2
#> 83  2016-03-17 04:30:00          44.22     2
#> 84  2016-03-17 04:32:00          48.84     2
#> 85  2016-03-17 04:34:00          45.38     2
#> 86  2016-03-17 04:36:00          42.38     2
#> 87  2016-03-17 04:38:00          46.53     2
#> 88  2016-03-17 04:40:00          53.45     2
#> 89  2016-03-17 04:42:00          55.53     2
#> 90  2016-03-17 04:44:00          48.38     2
#> 91  2016-03-17 04:46:00          45.84     2
#> 92  2016-03-17 04:48:00          49.53     2
#> 93  2016-03-17 04:50:00          37.53     2
#> 94  2016-03-17 04:52:00          36.61     2
#> 95  2016-03-17 04:54:00          37.99     2
#> 96  2016-03-17 04:56:00          39.38     2
#> 97  2016-03-17 04:58:00          39.15     2
#> 98  2016-03-17 05:00:00          42.38     2
#> 99  2016-03-17 05:02:00          47.92     2
#> 100 2016-03-17 05:04:00          46.53     2
#> 101 2016-03-17 05:06:00          43.76     2
#> 102 2016-03-17 05:06:00          43.76     2
#> 103 2016-03-17 05:08:00          43.76     2
#> 104 2016-03-17 05:10:00          44.22     2
#> 105 2016-03-17 05:10:00          44.22     2
#> 106 2016-03-17 05:12:00          43.99     2
#> 107 2016-03-17 05:14:00          49.07     2
#> 108 2016-03-17 05:16:00          51.38     2
#> 109 2016-03-17 05:18:00          43.99     2
#> 110 2016-03-17 05:20:00          50.92     2
#> 111 2016-03-17 05:20:00          50.92     2
#> 112 2016-03-17 05:22:00          55.07     2
#> 113 2016-03-17 05:24:00          57.15     2
#> 114 2016-03-17 05:26:00          56.92     2
#> 115 2016-03-17 05:28:00          57.15     2
#> 116 2016-03-17 05:30:00          57.15     2
#> 117 2016-03-17 05:32:00          56.69     2
#> 118 2016-03-17 05:34:00          56.23     2
#> 119 2016-03-17 05:36:00          49.30     2
#> 120 2016-03-17 05:38:00          47.22     2
#> 121 2016-03-17 05:40:00          51.84     2
#> 122 2016-03-17 05:42:00          54.15     2
#> 123 2016-03-17 05:44:00          55.07     2
#> 124 2016-03-17 05:46:00          59.00     2
#> 125 2016-03-17 05:48:00          58.07     2
#> 126 2016-03-17 05:50:00          52.07     2
#> 127 2016-03-17 05:52:00          49.99     2
#> 128 2016-03-17 05:54:00          51.15     2
#> 129 2016-03-17 05:56:00          55.30     2
#> 130 2016-03-17 05:58:00          61.07     2
#> 131 2016-03-17 06:00:00          65.00     2
#> 132 2016-03-17 06:02:00          68.23     2
#> 133 2016-03-17 06:04:00          71.70     2
#> 134 2016-03-17 06:06:00          76.09     2
#> 135 2016-03-17 06:08:00          79.79     2
#> 136 2016-03-17 06:10:00          83.48     2
#> 137 2016-03-17 06:12:00          87.18     2
#> 138 2016-03-17 06:14:00          91.57     2
#> 139 2016-03-17 06:16:00          95.04     2
#> 140 2016-03-17 06:18:00          97.12     2
#> 141 2016-03-17 06:20:00         101.29     2
#> 142 2016-03-17 06:22:00          99.21     2
#> 143 2016-03-17 06:24:00         101.06     2
#> 144 2016-03-17 06:26:00         103.60     2
#> 145 2016-03-17 06:28:00         103.60     2
#> 146 2016-03-17 06:30:00         104.06     2
#> 147 2016-03-17 06:32:00         103.83     2
#> 148 2016-03-17 06:34:00         102.91     2
#> 149 2016-03-17 06:36:00         100.59     2
#> 150 2016-03-17 06:38:00         100.59     2
#> 151 2016-03-17 06:40:00         100.59     2
#> 152 2016-03-17 06:42:00         100.59     2
#> 153 2016-03-17 06:44:00         100.36     2
#> 154 2016-03-17 06:46:00          96.66     2
#> 155 2016-03-17 06:48:00          92.50     2
#> 156 2016-03-17 06:50:00          88.34     2
#> 157 2016-03-17 06:52:00          83.02     2
#> 158 2016-03-17 06:54:00          78.40     2
#> 159 2016-03-17 06:56:00          84.18     2
#> 160 2016-03-17 06:58:00          92.73     2
#> 161 2016-03-17 07:00:00          98.74     2
#> 162 2016-03-17 07:02:00          99.90     2
#> 163 2016-03-17 07:04:00          99.90     2
#> 164 2016-03-17 07:06:00          99.90     2
#> 165 2016-03-17 07:08:00          99.21     2
#> 166 2016-03-17 07:10:00          95.04     2
#> 167 2016-03-17 07:12:00          92.50     2
#> 168 2016-03-17 07:14:00          92.73     2
#> 169 2016-03-17 07:16:00          92.50     2
#> 170 2016-03-17 07:18:00          92.73     2
#> 171 2016-03-17 07:20:00          92.73     2
#> 172 2016-03-17 07:22:00          92.27     2
#> 173 2016-03-17 07:24:00          83.25     2
#> 174 2016-03-17 07:26:00          80.71     2
#> 175 2016-03-17 07:28:00          85.80     2
#> 176 2016-03-17 07:30:00          85.80     2
#> 177 2016-03-17 07:32:00          85.80     2
#> 178 2016-03-17 07:34:00          85.80     2
#> 179 2016-03-17 07:36:00          85.80     2
#> 180 2016-03-17 07:38:00          85.80     2
#> 181 2016-03-17 07:40:00          91.34     2
#> 182 2016-03-17 07:42:00          91.57     2
#> 183 2016-03-17 07:44:00          98.28     2
#> 184 2016-03-17 07:46:00          99.67     2
#> 185 2016-03-17 07:48:00          96.20     2
#> 186 2016-03-17 07:50:00         101.75     2
#> 187 2016-03-17 07:52:00         101.29     2
#> 188 2016-03-17 07:54:00         100.59     2
#> 189 2016-03-17 07:56:00          97.36     2
#> 190 2016-03-17 07:58:00          91.81     2
#> 191 2016-03-17 08:00:00          90.19     2
#> 192 2016-03-17 08:02:00          87.41     2
#> 193 2016-03-17 08:04:00          89.49     2
#> 194 2016-03-17 08:06:00          96.89     2
#> 195 2016-03-17 08:08:00          84.64     2
#> 196 2016-03-17 08:10:00          80.25     2
#> 197 2016-03-17 08:12:00          86.72     2
#> 198 2016-03-17 08:14:00          89.03     2
#> 199 2016-03-17 08:16:00          88.57     2
#> 200 2016-03-17 08:18:00          88.57     2
#> 201 2016-03-17 08:20:00          82.33     2
#> 202 2016-03-17 08:22:00          71.93     2
#> 203 2016-03-17 08:24:00          66.62     2
#> 204 2016-03-17 08:26:00          64.31     2
#> 205 2016-03-17 08:28:00          63.85     2
#> 206 2016-03-17 08:30:00          61.77     2
#> 207 2016-03-17 08:32:00          63.85     2
#> 208 2016-03-17 08:34:00          61.07     2
#> 209 2016-03-17 08:36:00          51.38     2
#> 210 2016-03-17 08:38:00          46.76     2
#> 211 2016-03-17 08:40:00          37.30     2
#> 212 2016-03-17 08:42:00          37.07     2
#> 213 2016-03-17 08:44:00          38.22     2
#> 214 2016-03-17 08:46:00          36.38     2
#> 215 2016-03-17 08:48:00          39.61     2
#> 216 2016-03-17 08:50:00          37.30     3
#> 217 2016-03-17 08:50:00          37.30     3
#> 218 2016-03-17 08:52:00          39.38     3
#> 219 2016-03-17 08:52:00          39.38     3
#> 220 2016-03-17 08:54:00          39.38     3
#> 221 2016-03-17 08:54:00          39.38     3
#> 222 2016-03-17 08:56:00          44.92     3
#> 223 2016-03-17 08:58:00          45.38     3
#> 224 2016-03-17 08:58:00          45.38     3
#> 225 2016-03-17 09:00:00          45.15     3
#> 226 2016-03-17 09:00:00          45.15     3

#### Example (2): acdc_access_timesteps()
acdc_access_timesteps(acdc_simplify(dat_acdc))

#> acdc_simplify() implemented for type = 'acs'.

#> [1] 226

#### Example (3): acdc_access_maps()
acdc_access_maps(acdc_simplify(dat_acdc))

#> acdc_simplify() implemented for type = 'acs'.

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> [[2]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> [[3]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> [[4]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : NA, NA  (min, max)
#> 
#> 
#> [[5]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : NA, NA  (min, max)
#> 
#> 
#> [[6]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : NA, NA  (min, max)
#> 
#> 
#> [[7]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : NA, NA  (min, max)
#> 
#> 
#> [[8]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> [[9]]
#> class      : RasterLayer 
#> dimensions : 57, 74, 4218  (nrow, ncol, ncell)
#> resolution : 257, 463  (x, y)
#> extent     : 695492.1, 714510.1, 6246657, 6273048  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=29 +datum=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : layer 
#> values     : NA, NA  (min, max)
#> 
#>