"acdc_record" class

An S3 class that defines the object returned by an acoustic-centroid/depth-contour (AC/DC) algorithm (ac, dc or acdc), after simplification via acdc_simplify or directly from internal routines.

Value

A named list

An "acdc_record" class object is a named list with the following elements: `map', `record', `time', `args', `chunks' and `simplify'. The main output element is the `map' RasterLayer that shows where the individual could have spent more or less time over the duration of the movement time series. The `record' element records time-specific maps of the possible locations of the individual, and can be used to plot maps of specific time points or to produce animations (for the time steps specified by the save_record_spatial argument). The `time' element is a dataframe that defines the times of sequential stages in the algorithm's progression, providing a record of computation time. The `args' element is a named list of user inputs that record the parameters used to generate the outputs (if save_args = TRUE, otherwise the `args' element is NULL). The `chunks' element is a list with chunk-specific information that is generated if the algorithm is implemented chunk-wise and then simplified via acdc_simplify with the keep_chunks = TRUE argument. The simplify element is a logical value that defines whether or not the object was created from ac/dc/acdc and acdc_simplify or an internal routine. Below, more detail about the `map' and `record' elements is provided.

(A) `map'

This is a RasterLayer that defines the locations in which an individual could have been (or was not), given the movement time series. Each cell is a count of the number of times when the acoustic and/or archival data were compatible with the individual having been in that cell, given the algorithm (unless normalisation has been implemented via normalise = TRUE in the algorithm or in post-processing via acdc_simplify). This provides an overall measure of the locations in which an individual could have spent more or less time, but not where it was.

(B) `record'

This is a temporal record of the possible locations of each individual, with one element for each primary time step. For the AC* algorithms (ac and acdc), the `primary time step' refers to acoustic time steps; for the DC algorithm (dc), the `primary time step' refers to archival time steps. Each element contains (a) a dataframe with the information for that time step and (b) a list of spatial objects, one for each `secondary time step' that record spatial outputs of the algorithm, as specified by save_record_spatial (see below). For the AC* algorithms, `secondary time steps' are the time steps in-between acoustic observations; for the DC algorithm, primary and secondary time steps are the same.

(i) record$dat

For the AC* algorithms, record$dat contains the following information:

`timestep_cumulative' is a cumulative vector that defines the time step of each observation (across all acoustic and archival time steps up to the current time step);

`timestep_detection' defines the time step of the detection time series;

`timestep_archival' defines the time steps between the current and next acoustic detection (if applicable);

`receiver_1_id' and `receiver_2_id' define the receiver at which the individual was detected at the current and next acoustic time steps;

`receiver_1_timestamp' and `receiver_2_timestamp' define the time stamps of these detections;

`time_btw_dets' provides the time (s) between these acoustic detections;

`archival_timestamp' defines the time stamp of archival observations between acoustic detections (if applicable);

`archival_depth' defines the depth (m) of the individual each archival time step (if applicable);

For the DC algorithm, record$dat contains the archival data for the time step, as inputted, with the following additional columns:

`timestep_cumulative' is a cumulative integer that defines the time step;

`depth_lwr' is a number that defines the lower bound for the individual's possible depth on the bathymetry layer (bathy) given the depth error model (calc_depth_error);

`depth_upr' is a number that defines the upper bound for the individual's possible depth on the bathymetry layer (bathy) given the depth error model (calc_depth_error);

`availability' is a logical variable that defines whether or not there were any cells on the bathymetry layer (bathy) within the boundaries for the individual's depth at that time step;

(ii) record$spatial

If save_record_spatial > 0, then for those time steps specified by save_record_spatial, each element in `record' contains a `spatial' list, with one element for each secondary time stamp, that records spatial information pertaining to the possible locations of the individual at that time step. For all algorithms, each `spatial' list includes the following elements:

`map_timestep' is a RasterLayer of all the positions that the individual could have occupied at that time step, given the algorithm;

`map_cumulative' is a RasterLayer of the cumulative of the number of times when the movement data were compatible with the individual being in that cell, under the specified algorithm, from all previous time steps up to the current time step (i.e., the sum of `map_timestep' across all time steps from the first time step to the current time step) (unless normalise = TRUE in which case the interpretation differs);

For the AC* algorithms, this list also includes information on the acoustic centroids:

`centroid' is a SpatialPolygonsDataFrame that defines the acoustic centroid at that time step, which depends on the time since the previous detection, the time until the next detection, the spatial arrangement of receivers, the mobility parameter and any spatial masks (i.e., the coastline and specified boundaries);

If the receiver at which the individual was detected at the current and next time step differs, the `spatial' list also contains the following elements:

`centroid_2' is a SpatialPolygonsDataFrame that defines the boundaries of the individual's location from the perspective of the receiver at which the individual was next detected;

For save_record_spatial = 0L or elements outside of the specified input to save_record_spatial, `spatial' is simply an empty list.

Author

Edward Lavender