Calculate detection days

The function calculates the total number of days (termed `detection days') during which individuals were detected at passive acoustic telemetry receivers. To implement the function, a dataframe with passive acoustic telemetry detections of individuals at receivers must be supplied. Detection days can be calculated for all combinations of individuals/receivers in this dataframe, for all combinations of specified individuals and receivers, or for specific individual/receiver pairs. The function returns a dataframe of detection days for these individual/receiver combinations or a vector of detection days that is matched against another dataframe.

get_detection_days(
  acoustics,
  individual_id = NULL,
  receiver_id = NULL,
  type = 1L,
  match_to = NULL,
  ...
)

Arguments

acoustics
A dataframe that contains passive acoustic telemetry detection time series (see dat_acoustics for an example). This should contain the following columns: a vector of individual IDs, named `individual_id'; a vector of receiver IDs, named `receiver_id'; and a POSIXct vector of time stamps when detections were made, named `timestamp'.

individual_id
(optional) A vector of individuals for which to calculate detection days.

receiver_id
(optional) A vector of receivers for which to calculate detection days.

type
If both individual_id and receiver_id are specified, then type is an integer that defines whether or not to calculate detection days for (a) each individual/receiver pair (type = 1L) or (b) all combinations of individuals/receivers.

match_to
(optional) A dataframe against which to match detection days. This must contain the `individual_id' and `receiver_id' column, as in acoustics. If supplied, an integer vector of detection days for individual/receiver combinations, matched against the individuals/receivers in match_to, is returned (see also Value).

...
Additional arguments passed to pblapply.

Value

The function returns a dataframe with the detection days for all, or specified, combinations of individuals and receivers. Note that if acoustics only contains individuals/receivers that made detections, then this will only contain individuals/receivers for/at which detections were made. Alternatively, if match_to is supplied, a vector of detection days, matched against each individual/receiver observation in that dataframe, is returned.

See also

Author

Edward Lavender

Examples

#### Example (1): Detection days between all combinations
# ... of detected individuals and receivers with detections
dat <- get_detection_days(dat_acoustics)

#### Example (2) Detection days between all combinations of specified
#... individuals/receivers
dat <- get_detection_days(dat_acoustics,
                          individual_id = 25,
                          receiver_id = c(25, 21),
                          type = 2L)

#### Example (3) Detection days between specified individual/receiver pairs
dat <- get_detection_days(dat_acoustics,
                          individual_id = c(25, 26),
                          receiver_id = c(25, 21),
                          type = 1L)

#### Example (4) Match detection days to another dataframe
dat_acoustics$detection_days <- get_detection_days(dat_acoustics,
                                                   match_to = dat_acoustics,
                                                   type = 1L)

utils::head(dat_acoustics)

#>             timestamp individual_id transmitter_id receiver_id    receiver
#> 1 2016-03-17 01:49:12            25   A69-1303-555          26 VR2W-108163
#> 2 2016-03-17 01:51:40            25   A69-1303-555          26 VR2W-108163
#> 3 2016-03-17 01:52:56            25   A69-1303-555          26 VR2W-108163
#> 4 2016-03-17 01:54:19            25   A69-1303-555          26 VR2W-108163
#> 5 2016-03-17 01:57:11            25   A69-1303-555          26 VR2W-108163
#> 6 2016-03-17 01:58:07            25   A69-1303-555          26 VR2W-108163
#>   receiver_long receiver_lat receiver_depth detection_days
#> 1     -5.610733     56.37628             25             16
#> 2     -5.610733     56.37628             25             16
#> 3     -5.610733     56.37628             25             16
#> 4     -5.610733     56.37628             25             16
#> 5     -5.610733     56.37628             25             16
#> 6     -5.610733     56.37628             25             16