Calculate the distances between connected cells in a Raster*

This function calculates distances between connected cells in a raster, given (a) the planar distances between connected cells and (b) differences in elevation.

lcp_costs(surface, verbose = TRUE)

Arguments

surface
A raster for which to calculate distances. The surface must be planar (i.e., Universal Transverse Mercator projection) with units of metres in the x, y and z directions. The surface's resolution is taken to define the distance between connected cells in the x and y directions and must be the same in both cases (for surface's with unequal horizontal resolution, resample can be used to equalise resolution).

verbose
A logical function that defines whether or not to print messages to the console to relay function progress.

Value

The function returns a named list of sparse dsCMatrix-class matrices that define the distances (m) between connected cells under a rook's or bishop's movement (`dist_rook' and `dist_bishop'), the planar and vertical distances between connected cells (`dist_planar' and `dist_vertical') and the total distance between connected cells (`dist_total').

Details

This function was motivated by the need to determine the shortest paths between locations over the seabed for benthic animals (see lcp_over_surface). An animal's movement can be conceptualised as that of a queen on a chessboard, which can move, in eight directions around its current position, across a surface. Movements in the x and y direction are termed `rook's movement' and movements in the diagonal direction are termed `bishop's movement'.

Under this framework, the distance that an entity must travel between connected cells depends on the planar distances between cells and their differences in elevation. Planar distances (\(d_p\), m) depend on the movement type: under a rook's movement (i.e., horizontally or vertically), the distance (\(d_{p,r}\)) between connected cells is extracted from the raster's resolution (which is assumed to be identical in the x and y directions); under a bishop's movement (i.e., diagonally), the distance between connected cells \(d_{p,b}\) is given by Pythagoras' Theorem: \(d_{p,b} = \sqrt{(d_{p, r}^2 + d_{p, r}^2)}\). Vertical distances (\(d_v\), m) are simply the differences in height between cells. The total distance (\(d_t\)) between any two connected cells is a combination of these distances given by Pythagoras' Theorem: \(d_t = \sqrt{(d_p^2 + d_v^2)}\).

The function returns a warning produced by transition which is implemented to facilitate the definition of distances before shortest paths/distances are computed by either method: `In .TfromR(x, transitionFunction, directions, symm) : transition function gives negative values'. This warning arises because the height differences between connecting cells can be negative. It can be safely ignored.

See also

Author

Edward Lavender

Examples

# In this example, consider the distances between connected cells in the example
# ... 'dat_gebco' raster. For speed, we will focus on a subset of the area.
# ... Within this area, we need to regularise the resolution:
boundaries <- raster::extent(707884.6, 709884.6, 6253404, 6255404)
blank      <- raster::raster(boundaries, res = c(5, 5))
r          <- raster::resample(dat_gebco, blank)
# Implement algorithm
costs      <- lcp_costs(r)

#> flapper::lcp_costs() called (@ 2021-12-02 11:55:20)... 
#> ... Defining transition matrices... 

#> Warning: transition function gives negative values

#> ... Calculating distance matrices... 
#> ... Assembling LCP costs... 
#> ... flapper::lcp_costs() call completed (@ 2021-12-02 11:55:22) after ~0.04 minutes. 

# Examine outputs
utils::str(costs)

#> List of 5
#>  $ dist_rook    :Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. ..@ i       : int [1:319200] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ p       : int [1:160001] 0 0 1 2 3 4 5 6 7 8 ...
#>   .. ..@ Dim     : int [1:2] 160000 160000
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:319200] 5 5 5 5 5 5 5 5 5 5 ...
#>   .. ..@ uplo    : chr "U"
#>   .. ..@ factors : list()
#>  $ dist_bishop  :Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. ..@ i       : int [1:318402] 1 0 2 1 3 2 4 3 5 4 ...
#>   .. ..@ p       : int [1:160001] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..@ Dim     : int [1:2] 160000 160000
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:318402] 7.07 7.07 7.07 7.07 7.07 ...
#>   .. ..@ uplo    : chr "U"
#>   .. ..@ factors : list()
#>  $ dist_planar  :Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. ..@ i       : int [1:637602] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ p       : int [1:160001] 0 0 1 2 3 4 5 6 7 8 ...
#>   .. ..@ Dim     : int [1:2] 160000 160000
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:637602] 5 5 5 5 5 5 5 5 5 5 ...
#>   .. ..@ uplo    : chr "U"
#>   .. ..@ factors : list()
#>  $ dist_vertical:Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. ..@ i       : int [1:637602] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ p       : int [1:160001] 0 0 1 2 3 4 5 6 7 8 ...
#>   .. ..@ Dim     : int [1:2] 160000 160000
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:637602] -0.166 -0.166 -0.166 -0.166 -0.166 ...
#>   .. ..@ uplo    : chr "U"
#>   .. ..@ factors : list()
#>  $ dist_total   :Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. ..@ i       : int [1:637602] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ p       : int [1:160001] 0 0 1 2 3 4 5 6 7 8 ...
#>   .. ..@ Dim     : int [1:2] 160000 160000
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:637602] 5 5 5 5 5 ...
#>   .. ..@ uplo    : chr "U"
#>   .. ..@ factors : list()