Access the `history' element of a pf_archive-class object

This function accesses and simplifies the `history' list in a pf_archive-class object.

pf_access_history(archive, bathy = NULL)

Arguments

archive
A pf_archive-class object.

bathy
(optional) A raster that defines the grid across the area over which particle filtering was applied. If unsupplied, this is extracted from archive if available.

Value

The function returns a dataframe that defines, for each time step (`timestep'), particle samples (`cell_id'), associated probabilities (`cell_pr') and, if bathy is available, cell coordinates (`cell_x', `cell_y' and `cell_z').

Details

From the `history' element of a pf_archive-class object, this function extracts particle samples as a dataframe with columns for time steps, cell IDs, cell probabilities and coordinates (if bathy is available).

Author

Edward Lavender

Examples

pf_access_history(dat_dcpf_histories)

#>     timestep cell_id     cell_pr    cell_x    cell_y    cell_z
#> 1          1    2921 0.500000000 136.97954 136.97954 136.97954
#> 2          1    3159 0.900690015 137.04983 137.04983 137.04983
#> 3          1    3237 0.500000000 137.88798 137.88798 137.88798
#> 4          1    3243 0.924141820 133.69964 133.69964 133.69964
#> 5          1    3318 0.740227808 136.58845 136.58845 136.58845
#> 6          1    3320 0.962029408 135.27935 135.27935 135.27935
#> 7          1    3320 0.962029408 135.27935 135.27935 135.27935
#> 8          1    3396 0.150401435 137.20907 137.20907 137.20907
#> 9          1    3559 0.356595862 134.48285 134.48285 134.48285
#> 10         1    3722 0.068916511 134.01993 134.01993 134.01993
#> 11         2    3078 0.777299861 138.39286 138.39286 138.39286
#> 12         2    3079 0.740227808 137.60779 137.60779 137.60779
#> 13         2    3157 0.900690015 138.53295 138.53295 138.53295
#> 14         2    3243 0.068916511 133.69964 133.69964 133.69964
#> 15         2    3315 0.900690015 138.55210 138.55210 138.55210
#> 16         2    3475 0.620827238 137.08812 137.08812 137.08812
#> 17         2    3477 0.620827238 135.95305 135.95305 135.95305
#> 18         2    3478 0.461605227 135.38551 135.38551 135.38551
#> 19         2    3723 0.977022630 133.47766 133.47766 133.47766
#> 20         2    3961 0.150401435 136.28141 136.28141 136.28141
#> 21         3    3157 0.993307149 138.53295 138.53295 138.53295
#> 22         3    3166 0.740227808 133.69382 133.69382 133.69382
#> 23         3    3240 0.777299861 135.79381 135.79381 135.79381
#> 24         3    3317 0.924141820 137.24300 137.24300 137.24300
#> 25         3    3396 0.962029408 137.20907 137.20907 137.20907
#> 26         3    3402 0.900690015 133.54282 133.54282 133.54282
#> 27         3    3406 0.222700139 133.14011 133.14011 133.14011
#> 28         3    3956 0.222700139 139.05929 139.05929 139.05929
#> 29         3    3963 0.777299861 135.17026 135.17026 135.17026
#> 30         3    3963 0.777299861 135.17026 135.17026 135.17026
#> 31         4    3006 0.924141820 134.06296 134.06296 134.06296
#> 32         4    3238 0.962029408 137.18992 137.18992 137.18992
#> 33         4    3317 0.740227808 137.24300 137.24300 137.24300
#> 34         4    3398 0.812214802 135.98698 135.98698 135.98698
#> 35         4    3399 0.620827238 135.37594 135.37594 135.37594
#> 36         4    3565 0.900690015 131.98428 131.98428 131.98428
#> 37         4    3568 0.150401435 134.61187 134.61187 134.61187
#> 38         4    3641 0.356595862 133.98912 133.98912 133.98912
#> 39         4    3801 0.812214802 135.13527 135.13527 135.13527
#> 40         4    3887 0.461605227 134.05864 134.05864 134.05864
#> 41         5    2843 0.620827238 135.61930 135.61930 135.61930
#> 42         5    3005 0.977022630 133.75853 133.75853 133.75853
#> 43         5    3009 0.777299861 134.97625 134.97625 134.97625
#> 44         5    3075 0.620827238 140.74807 140.74807 140.74807
#> 45         5    3235 0.900690015 139.28409 139.28409 139.28409
#> 46         5    3400 0.977022630 134.76490 134.76490 134.76490
#> 47         5    3405 0.620827238 132.38837 132.38837 132.38837
#> 48         5    3719 0.500000000 135.64673 135.64673 135.64673
#> 49         5    3889 0.924141820 134.28312 134.28312 134.28312
#> 50         5    3963 0.356595862 135.17026 135.17026 135.17026
#> 51         6    3000 0.620827238 137.33717 137.33717 137.33717
#> 52         6    3008 0.977022630 134.67182 134.67182 134.67182
#> 53         6    3086 0.962029408 133.87839 133.87839 133.87839
#> 54         6    3317 0.900690015 137.24300 137.24300 137.24300
#> 55         6    3318 0.202010721 136.58845 136.58845 136.58845
#> 56         6    3329 0.500000000 135.31153 135.31153 135.31153
#> 57         6    3397 0.356595862 136.59803 136.59803 136.59803
#> 58         6    3637 0.900690015 136.14046 136.14046 136.14046
#> 59         6    3641 0.740227808 133.98912 133.98912 133.98912
#> 60         6    3886 0.740227808 133.94639 133.94639 133.94639
#> 61         7    2917 0.740227808 140.46788 140.46788 140.46788
#> 62         7    2921 0.962029408 136.97954 136.97954 136.97954
#> 63         7    3083 0.017742716 134.46750 134.46750 134.46750
#> 64         7    3089 0.777299861 135.06007 135.06007 135.06007
#> 65         7    3238 0.620827238 137.18992 137.18992 137.18992
#> 66         7    3247 0.740227808 134.08207 134.08207 134.08207
#> 67         7    3493 0.356595862 138.84399 138.84399 138.84399
#> 68         7    3560 0.740227808 133.94945 133.94945 133.94945
#> 69         7    3568 0.740227808 134.61187 134.61187 134.61187
#> 70         7    3635 0.924141820 137.21613 137.21613 137.21613
#> 71         8    3898 0.021065363 129.29371 129.29371 129.29371
#> 72         8    3898 0.021065363 129.29371 129.29371 129.29371
#> 73         8    3899 0.008469216 127.54157 127.54157 127.54157
#> 74         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 75         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 76         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 77         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 78         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 79         8    3977 0.017742716 129.34083 129.34083 129.34083
#> 80         8    3978 0.008469216 127.45340 127.45340 127.45340
#> 81         9    4063 0.068916511 115.49957 115.49957 115.49957
#> 82         9    4142 0.222700139 115.14083 115.14083 115.14083
#> 83         9    4142 0.424750760 115.14083 115.14083 115.14083
#> 84         9    4221 0.222700139 115.05267 115.05267 115.05267
#> 85         9    4221 0.222700139 115.05267 115.05267 115.05267
#> 86         9    4300 0.222700139 115.23507 115.23507 115.23507
#> 87         9    4301 0.092064003 112.80651 112.80651 112.80651
#> 88         9    4301 0.111940843 112.80651 112.80651 112.80651
#> 89         9    4301 0.111940843 112.80651 112.80651 112.80651
#> 90         9    4380 0.051872798 113.12419 113.12419 113.12419
#> 91        10    4303 0.900690015 107.94940 107.94940 107.94940
#> 92        10    4303 0.900690015 107.94940 107.94940 107.94940
#> 93        10    4304 0.461605227 105.52084 105.52084 105.52084
#> 94        10    4462 0.740227808 105.61507 105.61507 105.61507
#> 95        10    4462 0.900690015 105.61507 105.61507 105.61507
#> 96        10    4463 0.620827238 102.91595 102.91595 102.91595
#> 97        10    4545 0.222700139  94.73041  94.73041  94.73041
#> 98        10    4622 0.202010721 100.85219 100.85219 100.85219
#> 99        10    4783 0.051872798  92.84905  92.84905  92.84905
#> 100       10    4858 0.010774753 107.21004 107.21004 107.21004
#> 101       11    4228 0.202010721  96.41539  96.41539  96.41539
#> 102       11    4387 0.461605227  94.50041  94.50041  94.50041
#> 103       11    4389 0.356595862  88.67417  88.67417  88.67417
#> 104       11    4547 0.222700139  89.41252  89.41252  89.41252
#> 105       11    4548 0.222700139  86.75919  86.75919  86.75919
#> 106       11    4783 0.202010721  92.84905  92.84905  92.84905
#> 107       11    4861 0.740227808  97.08341  97.08341  97.08341
#> 108       11    4862 0.222700139  93.70787  93.70787  93.70787
#> 109       11    4862 0.962029408  93.70787  93.70787  93.70787
#> 110       11    4864 0.620827238  86.95678  86.95678  86.95678
#> 111       12    4389 0.900690015  88.67417  88.67417  88.67417
#> 112       12    4469 0.962029408  86.39001  86.39001  86.39001
#> 113       12    4548 0.993307149  86.75919  86.75919  86.75919
#> 114       12    4705 0.111940843  89.15582  89.15582  89.15582
#> 115       12    4707 0.461605227  84.32463  84.32463  84.32463
#> 116       12    4709 0.356595862  79.53754  79.53754  79.53754
#> 117       12    4943 0.962029408  87.81560  87.81560  87.81560
#> 118       12    5104 0.620827238  79.67062  79.67062  79.67062
#> 119       12    5106 0.620827238  74.12384  74.12384  74.12384
#> 120       12    5265 0.092064003  72.85097  72.85097  72.85097
#> 121       13    4552 0.500000000  76.14586  76.14586  76.14586
#> 122       13    4867 0.812214802  79.23674  79.23674  79.23674
#> 123       13    4871 0.812214802  70.70169  70.70169  70.70169
#> 124       13    4952 0.047249756  66.67345  66.67345  66.67345
#> 125       13    5026 0.977022630  76.02283  76.02283  76.02283
#> 126       13    5029 0.740227808  70.40090  70.40090  70.40090
#> 127       13    5185 0.962029408  74.43990  74.43990  74.43990
#> 128       13    5266 0.977022630  71.07040  71.07040  71.07040
#> 129       13    5345 0.740227808  71.26204  71.26204  71.26204
#> 130       13    5429 0.356595862  63.38557  63.38557  63.38557
#> 131       14    4636 0.461605227  63.11225  63.11225  63.11225
#> 132       14    4716 0.222700139  61.62966  61.62966  61.62966
#> 133       14    4956 0.202010721  57.18187  57.18187  57.18187
#> 134       14    5034 0.900690015  61.03102  61.03102  61.03102
#> 135       14    5036 0.461605227  55.69928  55.69928  55.69928
#> 136       14    5196 0.222700139  53.83588  53.83588  53.83588
#> 137       14    5273 0.461605227  59.40138  59.40138  59.40138
#> 138       14    5513 0.461605227  56.11490  56.11490  56.11490
#> 139       14    5590 0.900690015  59.28117  59.28117  59.28117
#> 140       14    5830 0.202010721  55.44025  55.44025  55.44025
#> 141       15    4717 0.977022630  57.48378  57.48378  57.48378
#> 142       15    4798 0.812214802  51.78779  51.78779  51.78779
#> 143       15    4880 0.356595862  41.81090  41.81090  41.81090
#> 144       15    4956 0.993307149  57.18187  57.18187  57.18187
#> 145       15    4958 0.620827238  48.68757  48.68757  48.68757
#> 146       15    5200 0.222700139  37.95396  37.95396  37.95396
#> 147       15    5280 0.461605227  37.89623  37.89623  37.89623
#> 148       15    5672 0.812214802  55.28290  55.28290  55.28290
#> 149       15    5752 0.900690015  54.12580  54.12580  54.12580
#> 150       15    6069 0.740227808  52.65030  52.65030  52.65030
#> 151       16    4560 0.500000000  48.28138  48.28138  48.28138
#> 152       16    4560 0.620827238  48.28138  48.28138  48.28138
#> 153       16    4719 0.962029408  49.19202  49.19202  49.19202
#> 154       16    4959 0.977022630  44.44043  44.44043  44.44043
#> 155       16    5119 0.461605227  42.17400  42.17400  42.17400
#> 156       16    5279 0.461605227  41.67489  41.67489  41.67489
#> 157       16    5360 0.150401435  37.83849  37.83849  37.83849
#> 158       16    5835 0.620827238  49.26138  49.26138  49.26138
#> 159       16    6314 0.016300297  44.29383  44.29383  44.29383
#> 160       16    6389 0.500000000  47.28266  47.28266  47.28266
#> 161       17    4880 0.500000000  41.81090  41.81090  41.81090
#> 162       17    5119 0.924141820  42.17400  42.17400  42.17400
#> 163       17    5279 0.993307149  41.67489  41.67489  41.67489
#> 164       17    5280 0.977022630  37.89623  37.89623  37.89623
#> 165       17    5360 0.962029408  37.83849  37.83849  37.83849
#> 166       17    5440 0.900690015  37.78075  37.78075  37.78075
#> 167       17    5519 0.900690015  40.92623  40.92623  40.92623
#> 168       17    6237 0.150401435  41.63400  41.63400  41.63400
#> 169       17    6240 0.068916511  37.20337  37.20337  37.20337
#> 170       17    6396 0.900690015  41.46088  41.46088  41.46088
#> 171       18    5200 0.962029408  37.95396  37.95396  37.95396
#> 172       18    5360 0.977022630  37.83849  37.83849  37.83849
#> 173       18    5439 0.924141820  41.17578  41.17578  41.17578
#> 174       18    5520 0.924141820  37.72301  37.72301  37.72301
#> 175       18    5599 0.900690015  40.67667  40.67667  40.67667
#> 176       18    6159 0.900690015  38.92980  38.92980  38.92980
#> 177       18    6238 0.977022630  40.15712  40.15712  40.15712
#> 178       18    6318 0.962029408  39.71575  39.71575  39.71575
#> 179       18    6319 0.900690015  38.43069  38.43069  38.43069
#> 180       18    6320 0.977022630  37.14563  37.14563  37.14563