List `history' files from a PF algorithm

This function creates an ordered vector (or list) of `history' files derived from the particle filtering (PF) algorithm (pf). This is applicable if pf is implemented with the write_history argument specified.

pf_access_history_files(
  root,
  use_absolute_paths = FALSE,
  use_list = FALSE,
  ...
)

Arguments

root
A string that defines the directory in which files are located.

use_absolute_paths
A logical variable that defines whether to return relative paths (FALSE) or absolute paths (TRUE) (see file_path_as_absolute).

use_list
A logical variable that defines whether or not return a vector (use_list = FALSE) or a list (use_list = TRUE).

...
Additional arguments passed to list.files (excluding full.names).

Value

The function returns an ordered list of file paths.

Details

This function requires the stringr package.

See also

Author

Edward Lavender

Examples

#### Example (1): Example with default arguments
# Define a directory in which to save files from PF
root <- paste0(tempdir(), "/pf/")
dir.create(root)

#> Warning: '/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf' already exists

# Implement the PF algorithm with write_history specified
# ... For speed, we will implement the algorithm using pre-defined data
pf_args <- dat_dcpf_histories$args
pf_args$calc_distance_euclid_fast <- TRUE
pf_args$write_history             <- list(file = root)
do.call(pf, pf_args)

#> flapper::pf() called (@ 2021-12-02 14:44:32)... 
#> ... Setting up function... 
#> ... Determining the set of possible starting locations (t = 1)... 
#> ... Implementing algorithm iteratively over time steps... 
#> ... ... Time = 1... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 2... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 3... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 4... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 5... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 6... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 7... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 8... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 9... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 10... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 11... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 12... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 13... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 14... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 15... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 16... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 17... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... ... ... For each particle, getting the possible positions for the next time step... 
#> ... ... Time = 18... 
#> ... ... ... Selecting candidate starting positions for the current time step... 
#> ... flapper::pf() call completed (@ 2021-12-02 14:44:33) after ~0.01 minutes. 

#> $history
#> $history[[1]]
#>    id_current pr_current timestep
#> 1        3243 0.92414182        1
#> 2        2921 0.50000000        1
#> 3        3396 0.15040143        1
#> 4        3159 0.90069001        1
#> 5        3320 0.96202941        1
#> 6        3237 0.50000000        1
#> 7        3318 0.74022781        1
#> 8        3559 0.35659586        1
#> 9        3722 0.06891651        1
#> 10       3320 0.96202941        1
#> 
#> $history[[2]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       3312  0.4616052        2
#> 2           NA          NA       3634  0.6208272        2
#> 3           NA          NA       3167  0.4616052        2
#> 4           NA          NA       3397  0.9770226        2
#> 5           NA          NA       3244  0.9770226        2
#> 6           NA          NA       3003  0.9006900        2
#> 7           NA          NA       2999  0.9241418        2
#> 8           NA          NA       3398  0.9770226        2
#> 9           NA          NA       2997  0.8122148        2
#> 10          NA          NA       3647  0.2020107        2
#> 
#> $history[[3]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       3557  0.9241418        3
#> 2           NA          NA       3250  0.7402278        3
#> 3           NA          NA       3089  0.9006900        3
#> 4           NA          NA       3713  0.9620294        3
#> 5           NA          NA       2835  0.8122148        3
#> 6           NA          NA       2844  0.9006900        3
#> 7           NA          NA       3470  0.8122148        3
#> 8           NA          NA       3951  0.2227001        3
#> 9           NA          NA       2923  0.9770226        3
#> 10          NA          NA       3395  0.9241418        3
#> 
#> $history[[4]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       2921  0.9241418        4
#> 2           NA          NA       3570  0.5000000        4
#> 3           NA          NA       2603  0.7402278        4
#> 4           NA          NA       3555  0.9241418        4
#> 5           NA          NA       2598  0.4247508        4
#> 6           NA          NA       2515  0.5000000        4
#> 7           NA          NA       3796  0.7402278        4
#> 8           NA          NA       4189  0.6208272        4
#> 9           NA          NA       3472  0.9241418        4
#> 10          NA          NA       3004  0.9620294        4
#> 
#> $history[[5]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       3244  0.7772999        5
#> 2           NA          NA       2355  0.9241418        5
#> 3           NA          NA       3626  0.0518728        5
#> 4           NA          NA       2360  0.6208272        5
#> 5           NA          NA       2522  0.9620294        5
#> 6           NA          NA       3325  0.4616052        5
#> 7           NA          NA       3477  0.9006900        5
#> 8           NA          NA       4508  0.4616052        5
#> 9           NA          NA       2839  0.9006900        5
#> 10          NA          NA       4432  0.4247508        5
#> 
#> $history[[6]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       4348  0.9241418        6
#> 2           NA          NA       2433  0.9006900        6
#> 3           NA          NA       2756  0.7402278        6
#> 4           NA          NA       2522  0.9933071        6
#> 5           NA          NA       4185  0.2227001        6
#> 6           NA          NA       2363  0.9006900        6
#> 7           NA          NA       3159  0.5000000        6
#> 8           NA          NA       3640  0.6208272        6
#> 9           NA          NA       2353  0.9241418        6
#> 10          NA          NA       3164  0.9770226        6
#> 
#> $history[[7]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       3877 0.42475076        7
#> 2           NA          NA       2444 0.96202941        7
#> 3           NA          NA       3078 0.96202941        7
#> 4           NA          NA       1963 0.22270014        7
#> 5           NA          NA       3485 0.46160523        7
#> 6           NA          NA       3316 0.62082724        7
#> 7           NA          NA       3002 0.81221480        7
#> 8           NA          NA       3561 0.96202941        7
#> 9           NA          NA       3887 0.01077475        7
#> 10          NA          NA       3959 0.46160523        7
#> 
#> $history[[8]]
#>    id_previous pr_previous id_current  pr_current timestep
#> 1           NA          NA       4372 0.008469216        8
#> 2           NA          NA       4293 0.008469216        8
#> 3           NA          NA       4372 0.008469216        8
#> 4           NA          NA       4372 0.008469216        8
#> 5           NA          NA       4372 0.008469216        8
#> 6           NA          NA       4293 0.008469216        8
#> 7           NA          NA       4293 0.008469216        8
#> 8           NA          NA       4293 0.008469216        8
#> 9           NA          NA       4293 0.008469216        8
#> 10          NA          NA       4293 0.008469216        8
#> 
#> $history[[9]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       4618 0.04724976        9
#> 2           NA          NA       4776 0.04724976        9
#> 3           NA          NA       4379 0.06891651        9
#> 4           NA          NA       4697 0.04724976        9
#> 5           NA          NA       4777 0.02106536        9
#> 6           NA          NA       4458 0.15040143        9
#> 7           NA          NA       4459 0.05187280        9
#> 8           NA          NA       4776 0.04724976        9
#> 9           NA          NA       4697 0.04724976        9
#> 10          NA          NA       4538 0.09206400        9
#> 
#> $history[[10]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       4939  0.8122148       10
#> 2           NA          NA       5099  0.3565959       10
#> 3           NA          NA       5098  0.4616052       10
#> 4           NA          NA       4621  0.8122148       10
#> 5           NA          NA       4699  0.9620294       10
#> 6           NA          NA       4304  0.2020107       10
#> 7           NA          NA       5100  0.2227001       10
#> 8           NA          NA       4383  0.5000000       10
#> 9           NA          NA       4701  0.7402278       10
#> 10          NA          NA       4700  0.9006900       10
#> 
#> $history[[11]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       4308 0.50000000       11
#> 2           NA          NA       5266 0.05187280       11
#> 3           NA          NA       5258 0.92414182       11
#> 4           NA          NA       4229 0.20201072       11
#> 5           NA          NA       5659 0.02297737       11
#> 6           NA          NA       4781 0.97702263       11
#> 7           NA          NA       5102 0.92414182       11
#> 8           NA          NA       4468 0.35659586       11
#> 9           NA          NA       4466 0.81221480       11
#> 10          NA          NA       5345 0.09206400       11
#> 
#> $history[[12]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5820  0.9006900       12
#> 2           NA          NA       4628  0.9241418       12
#> 3           NA          NA       5022  0.9770226       12
#> 4           NA          NA       4154  0.2020107       12
#> 5           NA          NA       5667  0.3565959       12
#> 6           NA          NA       5430  0.3565959       12
#> 7           NA          NA       4310  0.9620294       12
#> 8           NA          NA       5508  0.6208272       12
#> 9           NA          NA       5030  0.2227001       12
#> 10          NA          NA       5586  0.7402278       12
#> 
#> $history[[13]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5826  0.9006900       13
#> 2           NA          NA       5027  0.7772999       13
#> 3           NA          NA       4155  0.9770226       13
#> 4           NA          NA       5905  0.6208272       13
#> 5           NA          NA       5910  0.4247508       13
#> 6           NA          NA       5108  0.9006900       13
#> 7           NA          NA       5666  0.9770226       13
#> 8           NA          NA       5821  0.9770226       13
#> 9           NA          NA       3995  0.9006900       13
#> 10          NA          NA       4869  0.9006900       13
#> 
#> $history[[14]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5908 0.92414182       14
#> 2           NA          NA       5905 0.99330715       14
#> 3           NA          NA       5111 0.77729986       14
#> 4           NA          NA       5916 0.07585818       14
#> 5           NA          NA       6071 0.90069001       14
#> 6           NA          NA       5747 0.96202941       14
#> 7           NA          NA       6226 0.46160523       14
#> 8           NA          NA       5825 0.97702263       14
#> 9           NA          NA       6065 0.92414182       14
#> 10          NA          NA       5909 0.97702263       14
#> 
#> $history[[15]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5915  0.9770226       15
#> 2           NA          NA       5355  0.2227001       15
#> 3           NA          NA       6306  0.9770226       15
#> 4           NA          NA       6229  0.8122148       15
#> 5           NA          NA       6159  0.4247508       15
#> 6           NA          NA       6228  0.9241418       15
#> 7           NA          NA       6071  0.9933071       15
#> 8           NA          NA       5835  0.9620294       15
#> 9           NA          NA       5829  0.9770226       15
#> 10          NA          NA       6151  0.9770226       15
#> 
#> $history[[16]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       6390  0.9006900       16
#> 2           NA          NA       5994  0.9620294       16
#> 3           NA          NA       6398  0.7402278       16
#> 4           NA          NA       6312  0.9006900       16
#> 5           NA          NA       5197  0.8122148       16
#> 6           NA          NA       5917  0.9241418       16
#> 7           NA          NA       5998  0.9006900       16
#> 8           NA          NA       6079  0.9770226       16
#> 9           NA          NA       6158  0.9770226       16
#> 10          NA          NA       6158  0.9770226       16
#> 
#> $history[[17]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5840  0.8122148       17
#> 2           NA          NA       6239  0.9620294       17
#> 3           NA          NA       5920  0.9006900       17
#> 4           NA          NA       5839  0.9006900       17
#> 5           NA          NA       6078  0.9770226       17
#> 6           NA          NA       6078  0.9770226       17
#> 7           NA          NA       5919  0.9620294       17
#> 8           NA          NA       6397  0.9770226       17
#> 9           NA          NA       5520  0.2227001       17
#> 10          NA          NA       6159  0.9770226       17
#> 
#> $history[[18]]
#>    id_previous pr_previous id_current pr_current timestep
#> 1           NA          NA       5680  0.9241418       18
#> 2           NA          NA       5680  0.9241418       18
#> 3           NA          NA       5759  0.9770226       18
#> 4           NA          NA       6399  0.9241418       18
#> 5           NA          NA       6079  0.9770226       18
#> 6           NA          NA       5600  0.9770226       18
#> 7           NA          NA       5519  0.9770226       18
#> 8           NA          NA       5440  0.9770226       18
#> 9           NA          NA       6240  0.9770226       18
#> 10          NA          NA       6000  0.9770226       18
#> 
#> 
#> $method
#> [1] "pf"
#> 
#> $args
#> $args$record
#> $args$record[[1]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[2]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[3]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[4]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[5]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[6]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[7]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[8]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[9]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[10]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[11]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[12]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[13]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[14]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[15]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[16]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[17]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> $args$record[[18]]
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 1  (min, max)
#> 
#> 
#> 
#> $args$data
#> NULL
#> 
#> $args$origin
#>          [,1]    [,2]
#> [1,] 708886.3 6254404
#> 
#> $args$calc_distance
#> [1] "euclid"
#> 
#> $args$calc_distance_euclid_fast
#> [1] TRUE
#> 
#> $args$bathy
#> class      : RasterLayer 
#> dimensions : 80, 80, 6400  (nrow, ncol, ncell)
#> resolution : 25, 25  (x, y)
#> extent     : 707884.6, 709884.6, 6253404, 6255404  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 37.08789, 192.5775  (min, max)
#> 
#> 
#> $args$calc_distance_graph
#> NULL
#> 
#> $args$calc_movement_pr
#> function(distance,...) {
#>                  pr <- stats::plogis(5 + distance * -0.05)
#>                  pr[distance > 200] <- 0
#>                  return(pr)
#>                }
#> <bytecode: 0x7f965b1e90b0>
#> 
#> $args$calc_movement_pr_from_origin
#> function(distance,...) {
#>                  pr <- stats::plogis(5 + distance * -0.05)
#>                  pr[distance > 200] <- 0
#>                  return(pr)
#>                }
#> <bytecode: 0x7f965b1e3270>
#> 
#> $args$mobility
#> [1] 200
#> 
#> $args$mobility_from_origin
#> [1] 200
#> 
#> $args$n
#> [1] 10
#> 
#> $args$resample
#> NULL
#> 
#> $args$update_history
#> NULL
#> 
#> $args$update_history_from_time_step
#> [1] 1
#> 
#> $args$write_history
#> $args$write_history$file
#> [1] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf/"
#> 
#> 
#> $args$cl
#> NULL
#> 
#> $args$use_all_cores
#> [1] FALSE
#> 
#> $args$seed
#> [1] 5
#> 
#> $args$verbose
#> [1] TRUE
#> 
#> $args$optimisers
#> $use_raster_operations
#> [1] TRUE
#> 
#> $use_calc_distance_euclid_backend_grass
#> [1] FALSE
#> 
#> $use_grass_dir
#> NULL
#> 
#> attr(,"class")
#> [1] "list"         "pf_optimiser"
#> 
#> 
#> attr(,"class")
#> [1] "list"       "pf_archive"

# List the files
files <- pf_access_history_files(root)
utils::head(files)

#> [1] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_1.rds"
#> [2] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_2.rds"
#> [3] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_3.rds"
#> [4] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_4.rds"
#> [5] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_5.rds"
#> [6] "/var/folders/lx/dhz6yx2n2b7bg93hwz8t97zr0000gq/T//RtmpoGmbum/pf//pf_6.rds"