"pf_archive" class

An S3 class that defines the object returned by pf or pf plus pf_simplify with the return = "archive" argument.

Value

This object is a named list, with three elements, which records the particles sampled at each time step, the method and the parameters used in the call to pf:

history

A list of dataframes, one for each time step, that record particle samples. Each dataframe comprises n rows (one for each particle) and the following five columns:

id_previous. An integer that uniquely defines a previous location on the record rasters. Following implementation of pf, this is absent for the first time step. For subsequent time steps, this is NA if the fast Euclidean distances method has been used, which does not `remember' previous locations, but specified otherwise.) For pf_simplify, id_previous is NA for the fist time step, unless an origin has been specified; for subsequent time steps, id_previous is included.

pr_previous. A double that defines the movement probabilities associated with previous locations. Following implementation of pf, this is absent for the first time step. For subsequent time steps, this is NA if the fast Euclidean distances method has been used, which does not `remember' previous locations, but specified otherwise. For pf_simplify probabilities are included for all time steps.

id_current. An integer that uniquely defines each location on the record rasters.

pr_current. A double that defines the probability of movement into each cell.

timestep. An integer that defines each time step.

Following implementations of pf_simplify an additional column is included:

dist_current. A double that defines the distance between id_previous and id_current.

method

A character that defines whether or not history was derived directly from pf (method = "pf"), in which case history contains all of the particles sampled at each time step, or via pf plus pf_simplify with return = "archive" (method = "pf_simplify"), in which case history contains the subset of particles at each time step that were re-sampled at the next time step; for implementations with summarise_pr = TRUE, for particles that were sampled multiple times on a given time step, this only one sample with an adjusted probability score that accounts for the number of times that a location was sampled (see pf_simplify).

args

A named list that records the function arguments passed to pf. This is as inputted to pf, but with the `calc_distance_graph' elements added if unsupplied and applicable.

See also

Author

Edward Lavender